feat: add page with difficult words
