+++ src/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>RS-lang.Team</title>
+  <link rel="icon" href="./favicon.ico" />
+</head>
+<body class="body">  
+</body>
+</html>
