+++ src/js/render_home_page.ts
