+++ src/js/render_statistics.ts
