--- start-page/README.md
@@ -1,12 +1,0 @@
-# Empty project template
-Empty project with configured webpack, typescript and es-lint
-
-### Easy to use:
-+ Download project files
-+ Go to project root directory
-+ Run `npm i`
-
-### Pre-defined npm scripts:
-+ `npm run build` -> build project to the _dist_ folder. Es-lint will be running before build.
-+ `npm run dev` -> run webpack dev server and open browser
-+ `npm run lint` -> run es-lint
