--- start-page/src/index.html
@@ -1,157 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>RS-lang.Team</title>
-  <link rel="icon" href="./favicon.ico" />
-</head>
-<body class="body">
-<!--
-  <header class="header">
-    <h1 class="header-current"></h1>
-    <div class="burger"></div>
-    <nav class="header-nav">
-      <button class="nav-btn nav-start">Home</button>
-      <button class="nav-btn nav-textbook">TextBook</button>
-      <button class="nav-btn nav-games">Games</button>
-      <button class="nav-btn nav-statistics">Statistics</button>
-      <button class="nav-btn nav-developers">About App</button>   
-    </nav>
-    <button class="header-login _logged"></button>
-  </header>
-
-
-  <div class="main-wrapper">   
-    <main class="main">
-
-      <section class="home">
-        <div class="home-intro home-article">
-          <h2 class="home-intro_title">RS Lang</h2>
-          <div class="home-intro_subtitle">
-            RS Lang is an application designed to help you with mastering 
-            the English language. Regular lessons with this application 
-            will help you expand your language base. 
-            To get access to the full functionality of the application, 
-            we recommend that you go through a quick registration process. 
-          </div>
-        </div>
-
-        <div class="home-textbook home-article">
-          <h2 class="home-textbook_title">TextBook</h2>
-          <div class="home-textbook_subtitle">
-            With TextBook you can gradually learn words section by section, 
-            practice correct pronunciation, get acquainted with examples of word usage 
-            in different contexts.  
-          </div>
-        </div>
-
-        <div class="home-games home-article">
-          <h2 class="home-games_title">Word Games</h2>
-          <div class="home-games_subtitle">
-            Do you think learning English is boring? 
-            How about learning by playing? 
-            There are two games to choose from. 
-            Audio Challenge - try to guess a word by ear.
-            Sprint - become the fastest, guessing words visually. 
-          </div>
-        </div>
-
-        <div class="home-statistics home-article">
-          <h2 class="home-statistics_title">Statistics</h2>
-          <div class="home-statistics_subtitle">
-            Keep track of your learning progress 
-            in the special section Statistics
-          </div>
-        </div>
-      </section> 
-      
-      <section class="developers">
-        <h2 class="developers-title">Our team</h2>
-        <div class="developers-wrapper">
-          <div class="developer">
-            <h3 class="developer-subtitle">Alexey Zabashta</h3>
-            <p class="developer-role">Frontend developer</p>
-            <p class="developer-work">Created and configured a copy of the back-end, 
-              created a repository, wrote code for the Audiocall 
-              and Sprint games, statistics page.
-            </p>
-          </div>
-          <div class="developer">
-
-            <h3 class="developer-subtitle">Artur Zabashta</h3>
-            <p class="developer-role">Frontend developer</p>
-            <p class="developer-work">Designed layout of the application, made main page, 
-              login form, TextBook page and dictionary.
-            </p>
-          </div>
-        </div>
-      </section>
-
-      <section class="textbook">
-        <div class="textbook-wrapper">
-          <div class="textbook-pagination">
-            <button class="textbook-prev_btn">Prev</button>
-            <select name="textbook-select" id="">
-              <option value="1">Page 1</option>
-              <option value="2">Page 2</option>
-              <option value="3">Page 3</option>
-              <option value="4">Page 4</option>
-              <option value="5">Page 5</option>
-              <option value="6">Page 6</option>
-              <option value="7">Page 7</option>
-              <option value="8">Page 8</option>
-              <option value="9">Page 9</option>
-              <option value="10">Page 10</option>
-              <option value="11">Page 11</option>
-            </select>
-            <button class="textbook-nex_tbtn">Next</button>
-            <button class="textbook-audio_btn">Audio challenge</button>
-            <button class="textbook-sprint_btn">Sprint</button>
-          </div>
-          <div class="textbook-list">
-            <div class="textbook-item">
-              <p class="textbook-word"></p>
-              <p class="textbook-transcript"></p>
-              <p class="textbook-translate"></p>
-              <p class="textbook-word_audio"></p>
-              <p class="textbook-example"></p>
-              <p class="textbook-example_translate"></p>
-              <p class="textbook-meaning"></p>
-              <p class="textbook-meaning_translate"></p>
-              <p class="textbook-word_audio__meaning"></p>
-            </div>
-          </div>
-
-
-        </div>
-
-      </section>
-
-
-    </main>
-  </div>
-  <footer class="footer">    
-      <div class="footer-rsschool">
-        <a href="https://rs.school/js/" target="_Blank"
-          ><div class="footer-rsschool_svg"></div
-        ></a>
-      </div>      
-      <div class="footer-github">
-        <a href="https://github.com/ArturZabashta" target="_Blank">
-          Artur
-          <div class="my-github"></div>
-        </a>
-        <div class="footer-github_logo"></div>
-        <a href="https://github.com/AlexeyZabashta" target="_Blank">
-          Alexey
-          <div class="my-github"></div>
-        </a>
-      </div>
-      <div class="footer-copyright">© 2022</div>    
-  </footer>
--->
-  
-</body>
-</html>
